create port json encoders/decoders
This is a major departure from the intents of the upstream tool.  It still
creates json encoders and decoders, but for the raw data format that is
used by constructors in javascript protobuf instead.  This allows us to
pass message types between js and elm using ports, which in turn allows us
to call grpc-web RPCs in javascript through ports.