pkg/elm: use separate encoders for each oneof variant
Since oneofs technically use different field numbers in the encoding, the javascript
internal representation of protobuf types uses different indexes in the array for
those different oneof fields.  This means that for ports to work properly, we need
to encode a oneof to nil for indexes that do not match the field's current value's
number.

It's weird.